clarifying comment about the https/ssl issue
